refactor: replace printer handlers with printer service integration
This change removes the dedicated printer handlers and integrates the printer service directly into the Electron app. This streamlines the setup process and centralizes printer-related functionality.
